+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Add new items at the end of the relevant section under **Unreleased**.
 
 ### Additions
 
-- Add `validate` subcommand to generate validate reports
+- Add `validate` subcommand to generate validate reports ([#79])
+- Add `--pull-request` option to validate proposals in a pull request
 
 ### Changes
 
@@ -82,3 +83,4 @@ This changelog's format is based on [Keep a Changelog](https://keepachangelog.co
 [#72]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/72
 [#74]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/74
 [#79]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/79
+[#83]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/83